[#41] 주석 처리 및 파일 분리
