+++ B.READ/B.READ/Sources/Presentation/Search/SearchVOModel.swift
@@ -1,0 +1,17 @@
+//
+//  SearchVOModel.swift
+//  B.READ
+//
+//  Created by 김도연 on 5/18/25.
+//
+
+import SwiftUI
+
+struct BookVO {
+  let isbn: String
+  let coverImage: Image
+  let title: String
+  let author: String
+  let publisher: String
+  let publishedDate: Date
+}
